+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/github/Contributor.java
@@ -41,7 +41,7 @@ public class Contributor {
 	}
 	
 	public void setContribution(String role, int commits) {
-		if (!locked) {
+		if (!locked || role.startsWith("translator,")) {
 			contributions.put(role, commits);
 		}
 	}