+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/localization/Translators.java
@@ -47,6 +47,10 @@ public class Translators {
 		addTranslator("Nekomitsuki", "cs", true);
 		addTranslator("GGGEDR", "cs", true);
 		addTranslator("jakmanda05", "cs", true);
+		addTranslator("Aile-Minicraftcz", "cs", true);
+		addTranslator("Tengoblinekcz", "cs", true);
+		addTranslator("sirhCCC", "cs", true);
+		addTranslator("Thezerix", "cs", true);
 		
 		// Translators - Russian
 		addTranslator("SoSeDiK", "ru", false);